Rename builder scope interfaces
