+++ src/commonMain/kotlin/io/github/alexandrepiveteau/graphs/MutableNetworkScope.kt
@@ -1,0 +1,4 @@
+package io.github.alexandrepiveteau.graphs
+
+/** A [MutableGraphScope] for [Network]s. */
+public interface MutableNetworkScope : MutableGraphScope
