+++ src/commonMain/kotlin/io/github/alexandrepiveteau/graphs/MutableUndirectedGraphScope.kt
@@ -1,0 +1,12 @@
+package io.github.alexandrepiveteau.graphs
+
+/** A [MutableGraphScope] for [UndirectedGraph]s. */
+public interface MutableUndirectedGraphScope : MutableGraphScope {
+
+  /**
+   * Adds a new edge to the graph.
+   *
+   * @param edge the edge to add.
+   */
+  public fun addEdge(edge: Edge)
+}
