--- src/commonMain/kotlin/io/github/alexandrepiveteau/graphs/builder/NetworkBuilderScope.kt
@@ -1,6 +1,0 @@
-package io.github.alexandrepiveteau.graphs.builder
-
-import io.github.alexandrepiveteau.graphs.Network
-
-/** A [GraphBuilderScope] for [Network]s. */
-public interface NetworkBuilderScope : GraphBuilderScope
